scripts/GinanUI/requirements.txt
ruamel.yaml~=0.18.15
pandas~=2.3.3
PySide6~=6.10.0
plotly~=6.3.1
numpy~=2.3.3
statsmodels~=0.14.5
requests~=2.32.5
hatanaka~=2.8.1
unlzw3~=0.2.3
beautifulsoup4~=4.14.2